BackEnd/go-backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
pymongo
python-dotenv
